refactor: IThumbnail 분리
